General preparation to support USB thumb drives.
So far, https://github.com/carlk3/no-OS-FatFS-SD-SPI-RPi-Pico was referenced as-is. However that project directly links the FatFs code to the SPI SD card code. So when FatFs needs to be used with both an SPI SD card and a USB thumb drive, then FatFs needs to be decoupled from SPI SD card.
Therefore, the FatFs code is now used (as-is) from the (already present) TinyUSB project, as there's no official FatFs repo. The SPI SD card code is copied from https://github.com/carlk3/no-OS-FatFS-SD-SPI-RPi-Pico - and minimally modified to allow to add USB thumb drive support.